refactor: :recycle: Componetizar logs en varios archivos
